refactor(tests): dissolve phantom components dir and mirror effects tests
