+++ tests/unit/renderer/presentation/effects/capture.effect.test.ts
