+++ tests/unit/renderer/presentation/effects/controls-auto-hide.effect.test.ts
@@ -4,7 +4,7 @@
 
 import { describe, it, expect, beforeEach, afterEach, vi } from 'vitest';
 import { ControlsAutoHide } from '@renderer/presentation/effects/controls-auto-hide.effect';
-import { createCallbackMap } from '../../../../../factories/index.js';
+import { createCallbackMap } from '../../../../factories/index.js';
 
 describe('ControlsAutoHide', () => {
   let autoHide;
