+++ tests/unit/renderer/presentation/effects/toolbar-auto-hide.effect.test.ts
@@ -4,8 +4,8 @@
 
 import { describe, it, expect, beforeEach, afterEach, vi } from 'vitest';
 import { ToolbarAutoHide } from '@renderer/presentation/effects/toolbar-auto-hide.effect';
-import { installMissingMutationObserverMock } from '../../../../../support/mocks/browser-api.installers.js';
-import { createCallbackMap } from '../../../../../factories/index.js';
+import { installMissingMutationObserverMock } from '../../../../support/mocks/browser-api.installers.js';
+import { createCallbackMap } from '../../../../factories/index.js';
 
 describe('ToolbarAutoHide', () => {
   let autoHide;
