+++ tests/unit/renderer/presentation/features/streaming/streaming-controls.component.test.ts
@@ -5,7 +5,7 @@
 import { describe, it, expect, beforeEach, vi } from 'vitest';
 import { StreamingControlsComponent } from '@renderer/presentation/features/streaming/streaming-controls.component.js';
 import { StreamInfoStore } from '@renderer/presentation/state/stream-info.store.js';
-import { createStreamingControlsElementsMock, createUIBodyClassManagerMock, createEventBus } from '../../../../factories/index.js';
+import { createStreamingControlsElementsMock, createUIBodyClassManagerMock, createEventBus } from '../../../../../factories/index.js';
 import { EventChannels } from '@platform/events';
 
 describe('StreamingControlsComponent', () => {
