+++ tests/unit/renderer/presentation/shared/device-status.component.test.ts
