+++ tests/unit/renderer/presentation/shared/status-notification.component.test.ts
